Shooter pivot moved to its own folder to make things look nice
