+++ src/main/java/com/koibots/robot/subsystems/Subsystems.java
@@ -1,5 +1,6 @@
 package com.koibots.robot.subsystems;
 
+import com.koibots.robot.subsystems.ShooterPivot.ShooterPivot;
 import com.koibots.robot.subsystems.swerve.Swerve;
 
 import java.util.function.Supplier;
